requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0
aiohttp>=3.8.0
asyncio>=3.4.3

# Memory and storage
sqlite3
pickle5>=0.0.12
redis>=4.5.0

# Terminal and system interaction
psutil>=5.9.0
subprocess32>=3.5.4

# Code execution
docker>=6.1.0
jupyter>=1.0.0

# Web scraping and search
beautifulsoup4>=4.12.0
selenium>=4.10.0
google-api-python-client>=2.88.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 